+++ pkg/generator/key/key.go
@@ -16,3 +16,18 @@ var (
 	ConfigVersion = project.Name() + ".giantswarm.io/config-version"
 	AppLabel      = project.Name() + ".giantswarm.io/app"
 )
+
+func ToAppCR(v interface{}) (v1alpha1.App, error) {
+	if v == nil {
+		return v1alpha1.App{}, microerror.Maskf(wrongTypeError, "expected non-nil, got %#v", v)
+	}
+
+	p, ok := v.(*v1alpha1.App)
+	if !ok {
+		return v1alpha1.App{}, microerror.Maskf(wrongTypeError, "expected %T, got %T", p, v)
+	}
+
+	c := p.DeepCopy()
+
+	return *c, nil
+}
