+++ service/controller/key/error.go
@@ -1,0 +1,13 @@
+package key
+
+import (
+	"github.com/giantswarm/microerror"
+)
+
+var wrongTypeError = &microerror.Error{
+	Kind: "wrongTypeError",
+}
+
+func IsWrongType(err error) bool {
+	return microerror.Cause(err) == wrongTypeError
+}
