+++ service/controller/resource/values/error.go
@@ -21,11 +21,3 @@ var invalidConfigError = &microerror.Error{
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
-
-var wrongTypeError = &microerror.Error{
-	Kind: "wrongTypeError",
-}
-
-func IsWrongType(err error) bool {
-	return microerror.Cause(err) == wrongTypeError
-}
